Fail-safe initialization
Temporarily removed the "show saved" command.  Actually "show saved"
should be sent first because it could be sent manually, and we could not
distinguish it from the automatially sent command.  But it does not have
a reliable end marker in the output, and we therefore have to place it
at the end of our fail-safe initialization queue.