New board analysis strategy
Gibbon will now honor the post-Crawford (aka did-Crawford) flag in the
board state.  If one of the two players is 1-away and the flag is not
set, we have the Crawford game.

The cube has been turned if both may_double flags are unset and both
dice show zero.  We additionally check that this is not the Crawford
game although this is not needed on FIBS at least, where the may_double
flags are set during the post-Crawford game.