Option to terminate unlimited matches
Fix for Savannah #34985.  When we have a saved match of unlimited
length, we have to give the user an option to invite to a new match of
limited length.  This is the only way to terminate the saved match,
unless you want to wait for FIBS to discard the saved match.