swap dice in opening roll if necessary
The opening position contains a redundancy in our internal
representation.  If white is on move, the first die must be higher than
the second one.  If black is on move, the second die must be higher than
the first one.  The server doesn't know that and we have to correct it
for the opening roll.