import friends list
Due to an SQLite locking issue, the database code is in a debugging
state of mess right now.  Besides, the database and archive code has to
be re-written to be threadsafe.  That means no calls to Gdk, and
probably the only way to solve the locking problem s is to make the
database class a singleton, and protect all methods against concurrent
access with mutexes.