destroy main window after cleaning up
Connecting the "destroy" event of the main window directly to
gtk_main_quit() is dangerous.  The windows gets destroyed _before_ the
signal is fired.  That means that all objects owned by that window have
already been finalized.

In our case, the board object was indirectly owned by the window, and
was therefore finalized before the application itself.  A couple of
objects were already invalid at that point.  It would, of course, be
better to correct the reference count of these objects so that the order
of finalization does not matter but it is cleaner anyway to use the same
shutdown procedure independently of the way the user closed the
application.